requirements.txt
fastapi
uvicorn
pydantic
google-cloud-pubsub
python-dotenv
confluent-kafka
google-generativeai
